Show warning for unpaid merch pickup Fixes https://jira.magfest.net/browse/MAGDEV-531 by adding a warning to the merch pickup message -- the registration form already warns of unpaid attendees. Also moves all merch functions into their own site section so we can grant merch volunteers their own access group for use at-con.
