+++ uber/templates/merch_admin/multi_merch_pickup.html
@@ -70,7 +70,7 @@ Enter the badge numbers of everyone involved (blank fields will be ignored).
 <h2> Merch Picked up </h2>
 
 <div class="center">
-    <a href="../registration/merch">Return to main merch page</a>
+    <a href="../merch_admin/index">Return to main merch page</a>
 </div>
 
 {{ picker_upper.full_name }} has been marked as picking up the following merch:
